Implement gradient accumulation (#30)
* Implement gradient accumulation

This allows training with larger batches than GPU memory can store at
once. By default, no splitting of batches is performed. If
num_batch_splits is added to the training yml, the batch size is split
into sub-batches and the gradients are accumulated for these
sub-batches before performing one update step. This is equivalent to
training with the higher batch size in the first place.

As a result, this allows us to exactly match A0's training scheme which
used a batch size of 4096 and allows us to also match their LR of 0.2.

* Improve comments