Add some useful scripts
fixorder.py is useful when starting training for the first time and rescoring has messed with the timestamps on the files so they don't match the numerical order anymore.
initsplit.py is a super fast alternative to split.sh that works as a once off when setting up training for the first time.
rescore.sh runs rescorer in a loop - should take more parameters...
stage.sh watches file system events and moves files to a different directory (on the same file system) so that rescorer never tries to rescore a partially written file.